# hq/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zako3 HQ API",
    "description": "RESTful API for Zako3 platform - manage users, communities (taps), and bot settings",
    "version": "1.0.0",
    "contact": {
      "name": "Zako3 Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication provider"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "summary": "List settings",
        "tags": [
          "settings"
        ],
        "description": "Lists all settings for the specified kind and optional scope filter. Only accessible by admins or the resource owner.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "guild",
                "admin"
              ]
            },
            "in": "query",
            "name": "settingsKind",
            "required": true,
            "description": "The kind of settings to list"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "global",
                "guild",
                "user",
                "perGuildUser",
                "admin"
              ]
            },
            "in": "query",
            "name": "scopeType",
            "required": false,
            "description": "Filter by scope type"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "userId",
            "required": false,
            "description": "User ID for user settings"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "guildId",
            "required": false,
            "description": "Guild ID for guild settings"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "keyId": {
                            "type": "string",
                            "description": "Setting key identifier"
                          },
                          "value": {
                            "description": "Setting value"
                          },
                          "scope": {
                            "type": "object",
                            "properties": {
                              "kind": {
                                "type": "string",
                                "enum": [
                                  "user",
                                  "guild",
                                  "admin"
                                ]
                              },
                              "type": {
                                "type": "string"
                              },
                              "guildId": {
                                "type": "string"
                              },
                              "userId": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "kind",
                              "type"
                            ],
                            "additionalProperties": false,
                            "description": "Setting scope"
                          },
                          "isDefault": {
                            "type": "boolean",
                            "description": "Whether this is the default value"
                          }
                        },
                        "required": [
                          "keyId",
                          "value",
                          "scope",
                          "isDefault"
                        ],
                        "additionalProperties": false
                      },
                      "description": "List of setting entries"
                    }
                  },
                  "required": [
                    "entries"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.zako3.com",
      "description": "Production server"
    }
  ],
...
}